project-F1-Multimodal-RAG/requirements.txt
torch
transformers
chromadb
pillow
requests
openai
python-dotenv
numpy
